requirements.txt
azure-functions==1.21.1
azure-functions-durable==1.2.9
azure-storage-blob==12.23.1
pandas==2.2.3
python-dotenv==1.0.1
SQLAlchemy==2.0.35
pydantic==2.9.2
pyodbc==5.1.0
numpy==2.1.2
slack-sdk==3.33.4
git+https://github.com/rahul-08-11/pyzohocrm.git